+++ app/(protected)/dashboard/[boardid]/_components/create-task.tsx
@@ -41,7 +41,7 @@ import Image from 'next/image'
 import { useRouter } from 'next/navigation'
 import { ChangeEvent, KeyboardEvent, useEffect, useState } from 'react'
 import { toast } from 'sonner'
-import style from './modal.module.css'
+import style from '@/components/modal/_component/modal.module.css'
 import { ITaskCreateOpen } from './modal/modal-type'
 import { ModalHead } from './modal/modal-head'
 import { useRecoilState } from 'recoil'
